Fix the import task on very long key
I was importing translation files with `keys` longer 
than the 255 VARCHAR setted in the entity.

So the search was always returning false because searching with 
a long key something trimed at 255 chars.

That's not the most glorious fix, but the translations 
files I'm working with are not glorious neither. I think 
a 255 chars key is enougth, this PR is only about 
making the imported works with bad translation files.