Escape domain name in URL
I noticed that unescaped domain name may cause issues sometimes.

To be more concrete, I have my translations in messages+intl-icu.LANG.yaml file, but when I click on that file in the list, no results (translation strings) are shown, because what ends up in the domain filter box is actually not "messages+intl-icu", but "messages intl-icu". This little patch fixes that issue.

Note: I think presenting "messages" and "messages+intl-icu" as separate domains is the actual problem, which this patch does not address. I just wanted to pick the lowest hanging fruit and be done with it for now.