Update ImportTranslationsCommand.php
Fix deprecation raised on Symfony 6.3:

Method "Symfony\Component\Console\Command\Command::configure()" might add "void" as a native return type declaration in the future. Do the same in child class "Lexik\Bundle\TranslationBundle\Command\ImportTranslationsCommand" now to avoid errors or add an explicit @return annotation to suppress this message.

Note that the "void" return type have been introduced in PHP 8.1 so it's in line with the constraint in composer.json.